Round the converted value to 2 decimal places

+++ Currency Converter/Currency.swift
@@ -11,5 +11,5 @@ import Foundation
 struct Currency: Codable {
     let code: String
     let name: String
-    var rate: Decimal = 0
+    var rate: Double = 0
 }
